Solved the issue, let me explain:
So the issue was that the request you are sending to the backend is an empty object so theres nothing there so it gives you the error that password is less than 8. So basically i am sending an empty object but once i included name and made it not = to id, then it worked. The only thing that 8 character function is checking for is password